chore: BlockDomainLoader 를 인터페이스 작성
